Перенёс AuthRepository в отдельный модуль
