+++ app/src/main/java/ru/iteco/fmhandroid/api/AuthApi.kt
@@ -3,7 +3,7 @@ package ru.iteco.fmhandroid.api
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.POST
-import ru.iteco.fmhandroid.dto.AuthState
+import ru.iteco.fmh.model.AuthState
 import ru.iteco.fmhandroid.dto.LoginData
 
 interface AuthApi {
