--- app/src/main/java/ru/iteco/fmhandroid/dto/AuthState.kt
@@ -1,6 +1,0 @@
-package ru.iteco.fmhandroid.dto
-
-data class AuthState(
-    val accessToken: String,
-    val refreshToken: String
-)
