+++ app/src/main/java/ru/iteco/fmhandroid/repository/authRepository/AuthRepositoryModule.kt
@@ -4,6 +4,7 @@ import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
+import ru.iteco.fmh.data.AuthRepository
 
 @InstallIn(SingletonComponent::class)
 @Module
