+++ app/src/main/java/ru/iteco/fmhandroid/viewmodel/AuthViewModel.kt
@@ -5,11 +5,11 @@ import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.launch
+import ru.iteco.fmh.data.AuthRepository
 import ru.iteco.fmhandroid.auth.AppAuth
 import ru.iteco.fmhandroid.exceptions.AuthorizationException
 import ru.iteco.fmhandroid.exceptions.LostConnectException
 import ru.iteco.fmhandroid.exceptions.UnknownException
-import ru.iteco.fmhandroid.repository.authRepository.AuthRepository
 import ru.iteco.fmhandroid.repository.userRepository.UserRepository
 import javax.inject.Inject
