+++ model/src/main/kotlin/ru/iteco/fmh/model/AuthState.kt
@@ -1,0 +1,6 @@
+package ru.iteco.fmh.model
+
+data class AuthState(
+    val accessToken: String,
+    val refreshToken: String
+)
